Refactoring PreviewParameters into preview_parameter.py module
