+++ transcriptic/commands.py
@@ -14,7 +14,8 @@ from jinja2 import Environment, PackageLoader
 from os.path import isfile
 from transcriptic.english import AutoprotocolParser
 from transcriptic.config import Connection
-from transcriptic.util import iter_json, flatmap, ascii_encode, makedirs, PreviewParameters
+from transcriptic.util import iter_json, flatmap, ascii_encode, makedirs
+from transcriptic.preview_parameters import PreviewParameters
 from transcriptic import routes
 
 import sys
@@ -757,7 +758,8 @@ def launch(api, protocol, project, save_input, local, accept_quote, params, test
             try:
                 with click.open_file(save_input, 'w') as f:
                     f.write(
-                        json.dumps(dict(parameters=quick_launch["raw_inputs"]), indent=2)
+                        json.dumps(dict(parameters=quick_launch["raw_inputs"]),
+                                   indent=2)
                     )
             except Exception as e:
                 print_stderr("\nUnable to save inputs: %s" % str(e))
@@ -771,13 +773,6 @@ def launch(api, protocol, project, save_input, local, accept_quote, params, test
                 try:
                     with click.open_file(save_input, 'w') as f:
                         f.write(json.dumps(params, indent=2))
-                except Exception as e:
-                    print_stderr("\nUnable to save inputs: %s" % str(e))
-            elif test_inputs:
-                try:
-                    with click.open_file(save_input, 'w') as f:
-                        pp = TestParameters(params)
-                        f.write(json.dumps(pp.preview, indent=2))
                 except Exception as e:
                     print_stderr("\nUnable to save inputs: %s" % str(e))
         else:
@@ -877,8 +872,8 @@ def launch(api, protocol, project, save_input, local, accept_quote, params, test
                 pp = PreviewParameters({'parameters': quick_launch["raw_inputs"]})
                 f.write(json.dumps(pp.preview, indent=2))
         except Exception as e:
-            print_stderr("\nUnable to save preview inputs: %s" % str(e))
-        return None
+            print_stderr("\nUnable to save preview inputs due to not being able "
+                         "to process: %s %s" % type(e), str(e))
 
 
 def select_org(api, config, organization=None):
